refactor: split pr scope cli wrapper
