+++ .github/workflows/quality-gates.yml
@@ -49,7 +49,7 @@ jobs:
       - name: Determine PR scope
         id: scope
         if: github.event_name == 'pull_request' && steps.changes.outputs.relevant == 'true'
-        run: pnpm exec tsx scripts/ci/pr-scope.ts --base ${{ github.event.pull_request.base.sha }} --github-output
+        run: pnpm exec tsx scripts/ci/pr-scope.cli.ts --base ${{ github.event.pull_request.base.sha }} --github-output
 
       - name: Run lint gate
         if: github.event_name != 'pull_request' || steps.changes.outputs.relevant == 'true'
